+++ views/common/_calendarTypeItem.php
@@ -7,13 +7,19 @@
  */
 use humhub\libs\Html;
 use humhub\widgets\ModalButton;
+use yii\helpers\Url;
 
 /* @var $model \humhub\modules\calendar\models\CalendarEntryType */
 /* @var $contentContainer \humhub\modules\content\components\ContentContainerActiveRecord */
 
 
-$editUrl = $contentContainer->createUrl('/calendar/container-config/edit-type', ['id' => $model->id]);
-$deleteUrl = $contentContainer->createUrl('/calendar/container-config/delete-type', ['id' => $model->id]);
+if($contentContainer) {
+    $editUrl = $contentContainer->createUrl('/calendar/container-config/edit-type', ['id' => $model->id]);
+    $deleteUrl = $contentContainer->createUrl('/calendar/container-config/delete-type', ['id' => $model->id]);
+} else {
+    $editUrl = URL::to(['/calendar/config/edit-type', 'id' => $model->id]);
+    $deleteUrl = URL::to(['/calendar/config/delete-type', 'id' => $model->id]);
+}
 
 ?>
 <div class="media" style="margin-top:5px;">
@@ -22,13 +28,17 @@ $deleteUrl = $contentContainer->createUrl('/calendar/container-config/delete-typ
             <span class="input-group-addon">
                 <i style="display:inline-block;width:16px;height:16px;background-color: <?= Html::encode($model->color)?>"></i>
             </span>
-            <input class="form-control" value="<?= Html::encode($model->name) ?>" type="text" readonly>
+            <input class="form-control" value="<?= Html::encode($model->name) ?>" title="<?= Yii::t('CalendarModule.base', 'Event type color');?>" type="text" readonly>
             <span class="input-group-addon">
-                <?= ModalButton::primary()->load($editUrl)->icon('fa-pencil')->xs() ?>
-                <?= ModalButton::danger()->post($deleteUrl)->confirm(
+                <?php if($contentContainer && $model->contentcontainer_id == null) : ?>
+                    <small><?=  Yii::t('CalendarModule.config', '(global)') ?></small>
+                <?php else: ?>
+                    <?= ModalButton::primary()->load($editUrl)->icon('fa-pencil')->xs() ?>
+                    <?= ModalButton::danger()->post($deleteUrl)->confirm(
                         Yii::t('CalendarModule.config', '<strong>Confirm</strong> Deletion'),
                         Yii::t('CalendarModule.config', 'Do you really want to delte this event type?'),
                         Yii::t('CalendarModule.config', 'Delete'))->icon('fa-times')->xs() ?>
+                <?php endif; ?>
             </span>
         </div>
     </div>