+++ widgets/views/wallEntry.php
@@ -2,6 +2,7 @@
 
 use humhub\modules\calendar\widgets\EntryDate;
 use humhub\modules\calendar\widgets\EntryParticipants;
+use humhub\widgets\Button;
 use humhub\widgets\MarkdownView;
 use yii\helpers\Html;
 use humhub\modules\calendar\models\CalendarEntryParticipant;
@@ -14,9 +15,9 @@ use humhub\modules\calendar\models\CalendarEntry;
 $color = $calendarEntry->color ? $calendarEntry->color : $this->theme->variable('info');
 ?>
 
-<div class="media event" data-action-component="calendar.CalendarEntry" data-calendar-entry="<?= $calendarEntry->id ?>">
-    <div class="media-body" style="padding-left:10px; border-left: 3px solid <?= $color ?>">
-        <div class="clearfix">
+<div class="media event" style="" data-action-component="calendar.CalendarEntry" data-calendar-entry="<?= $calendarEntry->id ?>">
+    <div class="y" style="padding-left:10px; border-left: 3px solid <?= $color ?>">
+        <div class="media-body clearfix">
             <a href="<?= $calendarEntry->getUrl(); ?>" class="pull-left" style="margin-right: 10px">
                 <i class="fa fa-calendar colorDefault" style="font-size: 35px;"></i>
             </a>
@@ -29,27 +30,14 @@ $color = $calendarEntry->color ? $calendarEntry->color : $this->theme->variable(
                 <?= EntryDate::widget(['entry' => $calendarEntry]); ?>
             </h5>
         </div>
-        <div class="row">
-            <div class="col-md-12">
-                <?php if (!empty($calendarEntry->description)) : ?>
-                    <div <?= ($collapse) ? 'data-ui-show-more' : '' ?> data-read-more-text="<?= Yii::t('CalendarModule.views_entry_view', "Read full description...") ?>" style="overflow:hidden">
-                        <?= MarkdownView::widget(['markdown' => $calendarEntry->description]); ?>
-                    </div>
-                <?php endif; ?>
+        <?php if (!empty($calendarEntry->description)) : ?>
+            <div <?= ($collapse) ? 'data-ui-show-more' : '' ?> data-read-more-text="<?= Yii::t('CalendarModule.views_entry_view', "Read full description...") ?>" style="overflow:hidden">
+                <?= MarkdownView::widget(['markdown' => $calendarEntry->description]); ?>
             </div>
-        </div>
+        <?php endif; ?>
     </div>
 
     <?php if ($calendarEntry->isParticipationAllowed()) : ?>
-        <?php if(!$calendarEntry->closed) : ?>
-            <br>
-            <div class="row">
-                <div class="col-md-12">
-                    <?= EntryParticipants::widget(['calendarEntry' => $calendarEntry]); ?>
-                </div>
-            </div>
-        <?php endif; ?>
-
         <?php if(!empty($calendarEntry->participant_info) && $calendarEntry->isParticipant()) : ?>
             <br />
             <div class="row">
@@ -61,38 +49,36 @@ $color = $calendarEntry->color ? $calendarEntry->color : $this->theme->variable(
                 </div>
             </div>
         <?php endif; ?>
+
+        <?php if(!$calendarEntry->closed) : ?>
+            <br>
+            <div class="row">
+                <div class="col-md-12">
+                    <?= EntryParticipants::widget(['calendarEntry' => $calendarEntry]); ?>
+                </div>
+            </div>
+        <?php endif; ?>
     <?php endif; ?>
 
     <?php if ($calendarEntry->canRespond()): ?>
        <div class="row" style="padding-top:10px">
             <div class="col-md-12">
+                <?= Button::defaultType(Yii::t('CalendarModule.views_entry_view', "Attend"))->sm()
+                    ->icon($participantSate === CalendarEntryParticipant::PARTICIPATION_STATE_ACCEPTED ? 'fa-check' : null)
+                    ->action('calendar.respond', $contentContainer->createUrl('/calendar/entry/respond', ['type' => CalendarEntryParticipant::PARTICIPATION_STATE_ACCEPTED, 'id' => $calendarEntry->id]))?>
 
-                <?php
-                $cssState = ["", "", "", ""];
-                $cssState[$participantSate] = "disabled";
-                ?>
-                <button data-action-click="calendar.respond" class="btn btn-default btn-sm" data-ui-loader <?= $cssState[CalendarEntryParticipant::PARTICIPATION_STATE_ACCEPTED] ?>
-                        data-action-url="<?= $contentContainer->createUrl('/calendar/entry/respond', ['type' => CalendarEntryParticipant::PARTICIPATION_STATE_ACCEPTED, 'id' => $calendarEntry->id]) ?>">
-                    <?= $participantSate === CalendarEntryParticipant::PARTICIPATION_STATE_ACCEPTED ? '<i class="fa fa-check"></i>' : '' ?>
-                    <?= Yii::t('CalendarModule.views_entry_view', "Attend") ?>
-                </button>
 
                 <?php if($calendarEntry->allow_maybe) : ?>
-                    <button data-action-click="calendar.respond" class="btn btn-default btn-sm" data-ui-loader  <?= $cssState[CalendarEntryParticipant::PARTICIPATION_STATE_MAYBE] ?>
-                            data-action-url="<?= $contentContainer->createUrl('/calendar/entry/respond', ['type' => CalendarEntryParticipant::PARTICIPATION_STATE_MAYBE, 'id' => $calendarEntry->id]) ?>">
-                        <?= $participantSate === CalendarEntryParticipant::PARTICIPATION_STATE_MAYBE ? '<i class="fa fa-check"></i>' : ''?>
-                        <?= Yii::t('CalendarModule.views_entry_view', "Maybe") ?>
-                    </button>
+                    <?= Button::defaultType(Yii::t('CalendarModule.views_entry_view', "Maybe"))->sm()
+                        ->icon($participantSate === CalendarEntryParticipant::PARTICIPATION_STATE_MAYBE ? 'fa-check' : null)
+                        ->action('calendar.respond', $contentContainer->createUrl('/calendar/entry/respond', ['type' => CalendarEntryParticipant::PARTICIPATION_STATE_MAYBE, 'id' => $calendarEntry->id]))?>
                 <?php endif; ?>
 
                 <?php if($calendarEntry->allow_decline) : ?>
-                    <button data-action-click="calendar.respond" class="btn btn-default btn-sm" data-ui-loader  <?= $cssState[CalendarEntryParticipant::PARTICIPATION_STATE_DECLINED] ?>
-                            data-action-url="<?= $contentContainer->createUrl('/calendar/entry/respond', ['type' => CalendarEntryParticipant::PARTICIPATION_STATE_DECLINED, 'id' => $calendarEntry->id]) ?>">
-                        <?= $participantSate === CalendarEntryParticipant::PARTICIPATION_STATE_DECLINED ? '<i class="fa fa-check"></i>' : ''?>
-                        <?= Yii::t('CalendarModule.views_entry_view', "Decline") ?>
-                    </button>
+                    <?= Button::defaultType(Yii::t('CalendarModule.views_entry_view', "Decline"))->sm()
+                        ->icon($participantSate === CalendarEntryParticipant::PARTICIPATION_STATE_DECLINED ? 'fa-check' : null)
+                        ->action('calendar.respond', $contentContainer->createUrl('/calendar/entry/respond', ['type' => CalendarEntryParticipant::PARTICIPATION_STATE_DECLINED, 'id' => $calendarEntry->id]))?>
                 <?php endif;?>
-
             </div>
         </div>
     <?php endif; ?>